add title for missing pages
